LLMs/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
llama-index
openai
python-dotenv
rich
pydantic
chromadb
# pysqlite3-binary
streamlit
transformers
matplotlib
seaborn
scikit-learn
pysyllables
jinja2
tqdm
numpy
PyMuPDF
pandas
langchain_community